+++ demos/middleware/middleware.ts
@@ -1,0 +1,45 @@
+import { NextResponse } from 'next/server'
+import { NextFetchEvent, NextRequest } from 'next/server'
+
+export function middleware(request: NextRequest, ev: NextFetchEvent) {
+  let response
+  const {
+    nextUrl: { pathname },
+  } = request
+
+  if (pathname.startsWith('/cookies')) {
+    response = NextResponse.next()
+    response.cookies.set('netlifyCookie', 'true')
+    return response
+  }
+
+  if (pathname.startsWith('/shows')) {
+    if (pathname.startsWith('/shows/rewrite-absolute')) {
+      response = NextResponse.rewrite(new URL('/shows/100', request.url))
+      response.headers.set('x-modified-in-rewrite', 'true')
+    }
+    if (pathname.startsWith('/shows/rewrite-external')) {
+      response = NextResponse.rewrite('http://example.com/')
+      response.headers.set('x-modified-in-rewrite', 'true')
+    }
+    if (pathname.startsWith('/shows/rewriteme')) {
+      const url = request.nextUrl.clone()
+      url.pathname = '/shows/100'
+      response = NextResponse.rewrite(url)
+      response.headers.set('x-modified-in-rewrite', 'true')
+    }
+
+    if (!response) {
+      response = NextResponse.next()
+    }
+
+    if (pathname.startsWith('/shows/static')) {
+      response.headers.set('x-middleware-date', new Date().toISOString())
+    }
+
+    response.headers.set('x-modified-edge', 'true')
+    response.headers.set('x-is-deno', 'Deno' in globalThis ? 'true' : 'false')
+
+    return response
+  }
+}